Move THPStorage definitions out of `torch/csrc/generic`
