+++ torch/csrc/Storage.h
@@ -1,18 +1,22 @@
 #ifndef THP_STORAGE_INC
 #define THP_STORAGE_INC
-#include <torch/csrc/THConcat.h>
+
+#include <torch/csrc/Types.h>
 
 #define THPStorageStr "torch._UntypedStorage"
-#define THPStorage_(NAME) TH_CONCAT_2(THPStorage_,NAME)
-
-#define THPStorage_Check(obj) \
-    PyObject_IsInstance(obj, THPStorageClass)
-
-#define THPStorage_CData(obj)           (obj)->cdata
-
 #define THPStorageBaseStr "StorageBase"
 
-#include <torch/csrc/generic/Storage.h>
-#include <torch/csrc/THGenerateByteType.h>
+struct THPStorage {
+  PyObject_HEAD
+  c10::StorageImpl *cdata;
+};
+
+TORCH_PYTHON_API PyObject * THPStorage_New(c10::intrusive_ptr<c10::StorageImpl> ptr);
+extern PyObject *THPStorageClass;
+
+bool THPStorage_init(PyObject *module);
+void THPStorage_postInit(PyObject *module);
+
+extern PyTypeObject THPStorageType;
 
 #endif
