--- torch/csrc/StorageDefs.h
@@ -1,5 +1,0 @@
-#pragma once
-struct THPStorage {
-  PyObject_HEAD
-  c10::StorageImpl *cdata;
-};
